adjusted typo for stretched_image

+++ keras_segmentation/predict.py
@@ -206,7 +206,7 @@ def get_window_cutlines(seg_arr, seg_img, coords, window_height_adj, pixels_per_
     window_only = seg_img[window_ymin:window_ymax, window_xmin:window_xmax]
     # create a blank canvas np.zeros(window_height, seg_arr[1])
     print(f'get window cutlines window_only.shape: {window_only.shape}')
-    stretched_image = window_only.resize(window_only.shape[0], window_height_adj*ppi)
+    stretched_image = window_only.resize(window_only.shape[0], int(window_height_adj*ppi))
 
     # warp image to canvas
     # transform = cv2.getPerspectiveTransform(ordered_corners, dimensions)
